PDBR Update 0.9 - HUGE CONTENT UPDATE
PDBR Update 0.9 contains several changes, some of which include a reticle with hitmarkers, a damage sprite for cloakers, which also now have random spawning. New sounds have been added, such as unique gunshots for the SWAT enemies, reload sounds, and cloakers now have a melee attack. The player now has limited ammo, either reloaded automatically when empty, or manually when "R" is pressed. When you still have ammo and reload, you'll reload slightly faster, and will have an extra bullet in the chamber, for a total of 31.

NOTE: Currently there is a bug with this version of PDBR, where the reload text will overlay the "/30" while reloading. This is being worked on, PDBR Update 1.0 will have this fixed.