PDBR Update 0.9.4b FINAL VERSION
This is it people, the final version. Due to GMS2 having updated so that I only have 30 days on a trial account before I can't use the program, this, at least for now, will be where I conclude development for PAYDAY: Bag Run. It's been fun, and I hope to continue development in the future. I'm unsure if much has changed between this version and normal 0.9.4, but this is where I must finish my work. To everyone who has played this game through its development cycle, thank you so much, and I hope you enjoyed! <3